==> 02_minus_to_zero.ipynb <==
import csv
import os
import pickle

#import the datasets
with open("Pickles/Datasets.pkl", "rb") as l:
    datasets = pickle.load(l)




game_data = (r'game_data')


for dir in os.listdir(game_data): 
    loc = os.path.join(game_data, dir)
    #os.chdir(loc)                                        # setting the working directory to data folder causes chaos with checkpoints
    for file in os.listdir(loc):
        if file.endswith("postprocessed.csv"):            # only the preprocess files
            sucA = []
            sucB = []
            offset = 0
            with open(os.path.join(loc,file), 'r') as f:     # open each file
                reader = csv.reader(f, delimiter = ',')      # read the file
                with open(os.path.join(loc,'zero_'+ file), 'a+') as new_csv:  # open new csv
                    writer = csv.writer(new_csv)             # write in that new csv
                    for row in reader:
                        if row != []:                        # for every row in file (reader) that is not empty
                            if float(row[3]) < 0:                   # if success score < 0 we set it to 0
                                row[3] = 0
                            if float(row[-1]) < 0:
                                row[-1] = 0 
                        writer.writerow(row)                 # and write that row
