==> 06_Descriptives.ipynb <==
import pickle
import pandas as pd
import numpy as np
from statistics import mean, stdev, mode, median


with open("Pickles/AllDatapoints.pkl", "rb") as cdict:
    data = pickle.load(cdict)

with open("Pickles/Datasets.pkl", "rb") as cdict:
    datasets = pickle.load(cdict)
    
with open("Pickles/Minima.pkl", "rb") as cdict:
    Minima = pickle.load(cdict)

#### Success scores

success_scores = {}
for dataset in datasets:
    success_scores[dataset] = []
    for frame in range(len(data[dataset])):
        success_scores[dataset].append(data[dataset][frame][0])


success_scores


descriptives = pd.DataFrame(index= ['Min', 'Max', 'Mean', 'Mode', 'Median', 'SD', 'rel_spread', 'Frequency_0', 'total_length', 'Share_0'])

for dataset in datasets:
    descriptives[dataset] = np.nan
    values = []
    values.append(min(success_scores[dataset]))
    values.append(max(success_scores[dataset]))
    values.append(mean(success_scores[dataset]))
    values.append(mode(success_scores[dataset]))
    values.append(median(success_scores[dataset]))
    values.append(stdev(success_scores[dataset]))
    values.append(stdev(success_scores[dataset]) / max(success_scores[dataset]))
    values.append(len(np.where(np.array(success_scores[dataset]) == 0)[0]))
    values.append(len(success_scores[dataset]))
    values.append((len(np.where(np.array(success_scores[dataset]) == 0)[0])/len(success_scores[dataset])))
    descriptives[dataset] = values

descriptives.to_csv('Results/Descriptives.csv')

descriptives

np.where(np.array(success_scores[dataset]) == 0)[0]

len(np.where(np.array(success_scores[dataset]) == 0)[0])
