==> 06_Visualizations.ipynb <==
## Köln vs Wolfsburg

import numpy as np
import pickle
import matplotlib.pyplot as plt
from matplotlib.ticker import PercentFormatter
from scipy import stats
from statistics import mean, stdev

# import the scores
with open("Pickles/wolfsburg_scores.pkl", "rb") as cdict:
    wolfsburg_scores = pickle.load(cdict)
    
with open("Pickles/wolfsburg_scores_opp.pkl", "rb") as cdict:
    wolfsburg_scores_opp = pickle.load(cdict)
    
with open("Pickles/koeln_scores.pkl", "rb") as cdict:
    koeln_scores = pickle.load(cdict)
    
with open("Pickles/koeln_scores_opp.pkl", "rb") as cdict:
    koeln_scores_opp = pickle.load(cdict)
    
# import the datasets
with open("Pickles/Datasets.pkl", "rb") as l:
    datasets = pickle.load(l)


### Köln vs Wolfsburg

# '#4FB39F' & '#b34f63'

# '#ecff4c' & '#012210'
# '#eb2e00' & '#00BDEB'

for dataset in datasets:  # for every dataset (a&i)#
    mini = min([min(koeln_scores[dataset]), min(wolfsburg_scores[dataset])])
    maxi = max([max(koeln_scores[dataset]), max(wolfsburg_scores[dataset])])

    koeln_scores[dataset] = np.array(koeln_scores[dataset])            # list to array
    # histogram for Köln' scores, bins = min bis max, weights = for each value in x = 1 / number of values so that the total weight is 1 and we have a percentage histogram
    plt.hist(koeln_scores[dataset], bins=np.linspace(mini, maxi, 100), alpha=0.5, label='Köln', weights=np.ones(len(koeln_scores[dataset])) / len(koeln_scores[dataset]), color='#eb2e00')
    # histogram for Wolfsbrug on top of the first
    plt.hist(wolfsburg_scores[dataset],bins = np.linspace(mini, maxi, 100), alpha=0.5, label='Wolfsburg', weights=np.ones(len(wolfsburg_scores[dataset])) /len(wolfsburg_scores[dataset]), color = '#00BDEB')
    plt.gca().yaxis.set_major_formatter(PercentFormatter(1))
    plt.legend(loc='upper right')
    plt.title("Frequency distribution of Success Scores \n for Vfl Wolfsburg and 1.FC Köln: Dataset " + dataset)
    plt.xlabel("Success Score")
    plt.ylabel("Frequency")
    plt.savefig(''.join(['Distributionen/','KW_', dataset, '.png']),facecolor ="w",
            edgecolor ='w', bbox_inches='tight', dpi=500)
    plt.show()
    i += 1
    # bins=np.arange(koeln_scores[dataset].min(), koeln_scores[dataset].max() + 1

for dataset in datasets:  # for every dataset (a&i)#
    mini = min([min(koeln_scores[dataset]), min(wolfsburg_scores[dataset])])
    maxi = max([max(koeln_scores[dataset]), max(wolfsburg_scores[dataset])])

    koeln_scores[dataset] = np.array(koeln_scores[dataset])            # list to array
    # histogram for Köln' scores, bins = min bis max, weights = for each value in x = 1 / number of values so that the total weight is 1 and we have a percentage histogram
    plt.hist(koeln_scores[dataset], bins=np.linspace(mini, maxi, 100), alpha=0.5, label='Köln', weights=np.ones(len(koeln_scores[dataset])) / len(koeln_scores[dataset]), color='#eb2e00')
    # histogram for Wolfsbrug on top of the first
    plt.hist(wolfsburg_scores[dataset],bins = np.linspace(mini, maxi, 100), alpha=0.5, label='Wolfsburg', weights=np.ones(len(wolfsburg_scores[dataset])) /len(wolfsburg_scores[dataset]), color = '#00BDEB')
    plt.gca().yaxis.set_major_formatter(PercentFormatter(1))
    plt.legend(loc='upper right')
    plt.title("Frequency distribution of Success Scores \n for Vfl Wolfsburg and 1.FC Köln: Dataset " + dataset)
    plt.xlabel("Success Score")
    plt.ylabel("Frequency")
    plt.ylim(0,0.05)
    plt.savefig(''.join(['Distributionen/','KW_', dataset, '_5.png']),facecolor ="w",
            edgecolor ='w', bbox_inches='tight', dpi=500)
    plt.show()

### Köln's opponents vs Wolfsburg's opponents

for dataset in datasets:  # for every dataset (a&i)
    mini = min([min(koeln_scores_opp[dataset]), min(wolfsburg_scores_opp[dataset])])
    maxi = max([max(koeln_scores_opp[dataset]), max(wolfsburg_scores_opp[dataset])])
    koeln_scores_opp[dataset] = np.array(koeln_scores_opp[dataset])            # list to array
    # histogram for Köln' scores, bins = min bis max, weights = for each value in x = 1 / number of values so that the total weight is 1 and we have a percentage histogram
    plt.hist(koeln_scores_opp[dataset], bins=np.linspace(mini, maxi, 100), alpha=0.5, label="Köln's opponents", weights=np.ones(len(koeln_scores_opp[dataset])) / len(koeln_scores_opp[dataset]), color='#eb2e00')
    # histogram for Wolfsbrug on top of the first
    plt.hist(wolfsburg_scores_opp[dataset],bins=np.linspace(mini, maxi, 100), alpha=0.5, label="Wolfsburg's opponents", weights=np.ones(len(wolfsburg_scores_opp[dataset])) /len(wolfsburg_scores_opp[dataset]), color = '#00BDEB')
    plt.gca().yaxis.set_major_formatter(PercentFormatter(1))
    plt.legend(loc='upper right')
    plt.title("Frequency distribution of Success Scores \n for the opponents of Vfl Wolfsburg and 1.FC Köln: Dataset " + dataset)
    plt.xlabel("Success Score")
    plt.ylabel("Frequency")
    plt.savefig(''.join(['Distributionen/','Opp_', dataset, '.png']),facecolor ="w",
            edgecolor ='w', bbox_inches='tight', dpi=500)
    plt.show()


for dataset in datasets:  # for every dataset (a&i)
    mini = min([min(koeln_scores_opp[dataset]), min(wolfsburg_scores_opp[dataset])])
    maxi = max([max(koeln_scores_opp[dataset]), max(wolfsburg_scores_opp[dataset])])
    koeln_scores_opp[dataset] = np.array(koeln_scores_opp[dataset])            # list to array
    # histogram for Köln' scores, bins = min bis max, weights = for each value in x = 1 / number of values so that the total weight is 1 and we have a percentage histogram
    plt.hist(koeln_scores_opp[dataset], bins=np.linspace(mini, maxi, 100), alpha=0.5, label="Köln's opponents", weights=np.ones(len(koeln_scores_opp[dataset])) / len(koeln_scores_opp[dataset]), color='#eb2e00')
    # histogram for Wolfsbrug on top of the first
    plt.hist(wolfsburg_scores_opp[dataset],bins=np.linspace(mini, maxi, 100), alpha=0.5, label="Wolfsburg's opponents", weights=np.ones(len(wolfsburg_scores_opp[dataset])) /len(wolfsburg_scores_opp[dataset]), color = '#00BDEB')
    plt.gca().yaxis.set_major_formatter(PercentFormatter(1))
    plt.legend(loc='upper right')
    plt.title("Frequency distribution of Success Scores \n for the opponents of Vfl Wolfsburg and 1.FC Köln: Dataset " + dataset)
    plt.xlabel("Success Score")
    plt.ylabel("Frequency")
    plt.ylim(0, 0.05)
    plt.savefig(''.join(['Distributionen/','Opp_', dataset, '_5.png']),facecolor ="w",
            edgecolor ='w', bbox_inches='tight', dpi=500)
    plt.show()


with open('Pickles/Korrelation.pkl', 'rb') as cdict:
    Korrelation = pickle.load(cdict)

help_df = pd.concat([goals_8.head(1), goals_8.tail(1)])

fig, axes = plt.subplots(nrows=1, ncols=1)
axes.scatter(x= goals_8['Group'], y= goals_8['30m_300s'], c='#005270')
axes.plot(goals_8['Group'], goals_8['30m_300s'], c='#005270', linestyle='dotted')
#axes.plot(help_df['Group'], help_df['30m_300s'], c= '#ff6600')
axes.text(7, 2, "".join(["$r_s$ = ", str(round(Korrelation['30m_300s'][0],2)), "\n", "p = ", str(round(Korrelation['30m_300s'][1],2)).lstrip('0'), "*"]))
plt.ylim(0, 12)
plt.xlabel("Group")
plt.ylabel("Number of Goals")
#plt.title('Success Score-Goal-Correlation')
#rect = plt.Rectangle(
    # (lower-left corner), width, height
 #   (0, 0), 0.98, 0.98, fill=False, color="k", lw=1, 
  #  zorder=100, transform=fig.transFigure, figure=fig)
#fig.patches.extend([rect])

plt.savefig('Korrelation_500dpi.png',facecolor ="w",
            edgecolor ='w', bbox_inches='tight', dpi=500)
    
plt.show()
